apps/api/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud
google-cloud-storage==2.14.0
google-auth==2.25.2

# ML Framework Support (install based on model type)
scikit-learn==1.3.2  # For sklearn models
# torch==2.1.0+cpu  # Uncomment for PyTorch models
# tensorflow==2.15.0  # Uncomment for TensorFlow models
# onnxruntime==1.16.3  # Uncomment for ONNX models

# Utilities
python-multipart==0.0.6  # For file uploads
pillow==10.1.0  # For image processing
numpy==1.26.2
aiofiles==23.2.1

# Monitoring & Logging
python-json-logger==2.0.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
